Fixes and changes to make docker experience more stable
Increased Exec timeout to allow docker to complete.
Added clean up script to remove partials from brew/downloads to workaround the curl: (22) The requested URL returned error: 416 error
Removed the shutdown line for the hello-docker scritp. The Hello-world container starts and terminates almost immediately (as it doesn’t have a long running process within it), so the shutdown is not required.
Added a new nginx container (that has a long running process) as a sample of how to start and stop a docker container