Basic changes for QNX support
1. pom.xml: Commenting some plugins because it was failing even before I
made my changes.

2. Kernel.java: The JamaicaVM does implement its own classloader, but
that class loader follows the URLClassLoader protocol. The problem with
FastClasspathScanner, however, is that it *doesn't* use the class loader
for speed/performance reasons to avoid class verification and resolving.

Quick workaround is to hardcode the Evergreen services but I need to
implement a better solution. An idea is to get the information in
build time instead of runtime time but it will not work to scan the
servcies located in the plugin folder. More info:
https://github.com/classgraph/classgraph/wiki/Build-Time-Scanning

3. Exec.java: Removing some paths that are not present in QNX and
modifying the way to get PATH.