Update IotCloudHelper::sendHttpRequest to send body
Updating the sendHttpRequest function to accept a byte array which will be sent as body of the http request.
Also moved the use of "HttpClientConnection conn" into a try resource block.
Added a unit test for the case when the body is provided.

I want to use this in the DCM Service to send a request to the GCM cloud service.

Test results:
[INFO] Results:
[INFO]
[INFO] Tests run: 243, Failures: 0, Errors: 0, Skipped: 0
[INFO]
[INFO]
[INFO] --- jacoco-maven-plugin:0.8.5:report (post-unit-test) @ evergreen-kernel ---
[INFO] Loading execution data file /Users/shsush/ninj/EG/aws-greengrass-kernel/target/jacoco.exec
[INFO] Analyzed bundle 'evergreen-kernel' with 106 classes
[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  03:29 min
[INFO] Finished at: 2020-07-14T11:18:22-07:00
[INFO] ------------------------------------------------------------------------