Set the owner of artifacts for a service when commands are first run.
Refactored generation of the user/group to run a service with.

Artifacts are set to the user/group running the service.
The run as user is set for a component in the deployment. If none is
provided, the default user/group is used. If no default is set, and
the kernel user is not root, the kernel user is used.

Artifacts that are not "run" are left unchanged - they are owned by
the kernel user.