Set artifact and artifact dir permissions
Permissions will be set to world readable for artifacts - this is temporary until the component model change is consumed

The artifacts directories need to be globally readable and executable so any user can enter the directories - otherwise sharing artifacts does not work

Ownership of artifacts directories is set to the nucleus user. Ownership of artifact files is set to the user running the component (or default user)

Update Exec to remove some environment variables for HOME, USER, that do not and should not be set for components. These took the values of the user running the kernel (e.g. root)

The `cd()` method to change to the user home directory is also dangerous and unused so it has been removed. We do not want to start processes in the user home directory.