Restrict loading of external plugins to cloud based recipes
With this change, only recipes which are from cloud are trusted and
loaded as plugins. This aims to create development friction for customers as
they cannot create a local plugin recipe and load it. We want to do this
for 2 reasons. 1. Our plugin interface is not final and is subject to changes
with later greengrass releases. 2. A malicious user cannot modify existing plugin
on the disk or load a new one.