Set the owner of artifacts for a service when commands are first run. (#555)
* Set the owner of artifacts for a service when commands are first run.

Refactored generation of the user/group to run a service with.

Artifacts are set to the user/group running the service.
The run as user is set for a component in the deployment. If none is
provided, the default user/group is used. If no default is set, and
the kernel user is not root, the kernel user is used.

Artifacts that are not "run" are left unchanged - they are owned by
the kernel user.

Set artifact and artifact dir permissions

Permissions will be set to world readable for artifacts - this is temporary until the component model change is consumed

The artifacts directories need to be globally readable and executable so any user can enter the directories - otherwise sharing artifacts does not work

Ownership of artifacts directories is set to the nucleus user. Ownership of artifact files is set to the user running the component (or default user)

Fixes so tests run as non-root user

Add default user and group to easy setup