Reduce maxInFlightPublishes to 1 (#792)
Shadow update logic during deployment relies on MQTT messages being sent
in order. With maxInFlightPublishes > 1 it is possible for shadow update
messages to reach the cloud out of order, resulting in a deployment that
appears to not have finished.

As a workaround, this change temporarily reduces maxInFlightPublishes and
removes the configuration so that customers cannot break themselves.