fix(cancellation): handle deployment cancellation correctly when deployment is not registered (#799)
Correctly handle the stage deployment is in when evaluating if it can be cancelled. There was a regression introduced in device deployment handling which causes the evaluation to not work in cases when deployment is not yet registered in the UpdateSystemPolicyService's pendingActions queue, i.e. while deployment is still in dependency resolution/download artifacts stage.