feat(fleetstatusservice): make periodic update interval for Fleet Status Service configurable
If the configured value is less than the default (86400 seconds),
then the periodic update interval for FSS is set to the default.
If the configured value is more than the default,
then the periodic update interval for FSS is set to that configured value.
Add a new system configuration node called fleetStatus to add all cofiguration needed by FSS.