fix: allow environment variables to override defaults in the loader
Use := to allow environment variables to override the defaults for GG_ROOT,
LOADER_DIR and CONFIG_FILE. This allows the caller to set a GG_ROOT
somewhere other than ../../../../ but still use the loader, config and jar.
This is useful to allow devices that use signed read-only root filesystems
to ship a loader and bootstrap the nucleus on a different persistent
filesystem on target.